chore: move cross validation into SupervisorClient
